+++ src/models/ValidatorInterface.ts
@@ -36,6 +36,7 @@ export type ValidationPlugin  = {
 };
 
 export interface ValidationPlugins {
+  [key: string]: ValidationPlugin | undefined;
   vjf?: ValidationPlugin;
   dvr?: ValidationPlugin;
   svk?: ValidationPlugin;